Code improvement: Addition of Response class to represent results without model

+++ Code/OnLineStore/OnLineStore.Core/BusinessLayer/Responses/IPagedResponse.cs
@@ -4,6 +4,6 @@
     {
         int ItemsCount { get; set; }
 
-        int PageCount { get; }
+        double PageCount { get; }
     }
 }
